Added scrolling, move out of navbar
Added magic number, removed jqueryi
